+++ backend/routes/users.py
@@ -7,8 +7,15 @@ from schemas.user_schemas import UserCreate, UserResponse
 from utils.auth import hash_password
 from utils.auth import verify_password, create_access_token
 from datetime import timedelta
+from utils.dependencies import get_current_user
+from models.User import User as UserModel
 
 router = APIRouter()
+
+def require_superuser(user: UserModel = Depends(get_current_user)):
+    if user.role != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Superuser access required")
+    return user
 
 @router.post("/user/register")
 def create_user(user: UserCreate, db: Session = Depends(get_db)):
@@ -33,7 +40,8 @@ def create_user(user: UserCreate, db: Session = Depends(get_db)):
         "user": {
             "id": new_user.id,
             "email": new_user.email
-        }
+        },
+        "role": "tech"
     }
 
 
@@ -55,16 +63,19 @@ def login_user(user: UserCreate, db: Session = Depends(get_db)):
         "token_type": "bearer",
         "user": {
             "id": db_user.id,
-            "email": db_user.email
+            "email": db_user.email,
+            "role": db_user.role
         }
     }
 
 
 #Get all users
 @router.get("/users", response_model=List[UserResponse])
-def get_users(db: Session = Depends(get_db)):
+def get_users(
+    db: Session = Depends(get_db),
+    user: UserModel = Depends(require_superuser)
+):
     return db.query(UserModel).all()
-
 
 #Get a single user
 @router.get("/user/{user_id}", response_model=UserResponse)
@@ -77,16 +88,27 @@ def get_user(user_id: int, db: Session = Depends(get_db)):
 
 #Update a user
 @router.patch("/user/{user_id}", response_model=UserResponse)
-def update_user(user_id: int, user: UserCreate, db: Session = Depends(get_db)):
+def update_user(
+    user_id: int,
+    user: UserCreate,
+    db: Session = Depends(get_db),
+    current_user: UserModel = Depends(get_current_user),
+):
     db_user = db.query(UserModel).filter(UserModel.id == user_id).first()
     if not db_user:
-        raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail = "User not found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User not found")
+    
     db_user.email = user.email
     if user.password:
         db_user.password = hash_password(user.password)
+
+    if user.role and current_user.role == "admin":
+        db_user.role = user.role
+
     db.commit()
     db.refresh(db_user)
     return db_user
+
 
 #Delete a user
 @router.delete("/user/{user_id}", response_model=UserResponse)