+++ backend/utils/dependencies.py
@@ -8,18 +8,15 @@ from utils.auth import SECRET_KEY, ALGORITHM
 
 bearer_scheme = HTTPBearer(auto_error=False) 
 def get_current_user(credentials: HTTPAuthorizationCredentials = Depends(bearer_scheme), db: Session = Depends(get_db)):
-    print("🔑 Credentials:", credentials)
     if not credentials:
         print("❌ No credentials")
         raise HTTPException(status_code=403)
 
     if credentials.scheme.lower() != "bearer":
-        print("❌ Wrong auth scheme:", credentials.scheme)
         raise HTTPException(status_code=403)
-#Ste
+    
     try:
         payload = jwt.decode(credentials.credentials, SECRET_KEY, algorithms=[ALGORITHM])
-        print("🧠 Token payload:", payload)
         email = payload.get("sub")
         if not email:
             print("❌ No sub in payload")